requirements.txt
pyzmq
PyQt6
